+++ dnd/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "iosfwd": "cpp",
+        "vector": "c"
+    }
+}
