+++ dnd/CMakeLists copy 2.txt
@@ -1,0 +1,2 @@
+ADD_SUBDIRECTORY(examples)
+ADD_SUBDIRECTORY(homework)
