+++ dnd/CMakeLists copy.txt
@@ -1,0 +1,8 @@
+ADD_SUBDIRECTORY(01_data_types)
+ADD_SUBDIRECTORY(02_expressions)
+ADD_SUBDIRECTORY(03_decisions)
+ADD_SUBDIRECTORY(04_repetition)
+ADD_SUBDIRECTORY(05_functions)
+ADD_SUBDIRECTORY(06_tic_tac_toe)
+ADD_SUBDIRECTORY(07_arrays)
+ADD_SUBDIRECTORY(08_memory)
