+++ dnd/README.md
@@ -1,0 +1,4 @@
+[![Open in Visual Studio Code](https://classroom.github.com/assets/open-in-vscode-f059dc9a6f8d3a56e377f745f24479a46679e63a5d9fe6f495e02850cd0d8118.svg)](https://classroom.github.com/online_ide?assignment_repo_id=5461171&assignment_repo_type=AssignmentRepo)
+# acc-cosc-1337-template
+C++ Development examples and homeworks starter code.  Includes GitHub Actions for building and unit testing C++ programs.
+Testing
