+++ dnd/dnd.cpp
@@ -5,17 +5,39 @@ using std::cout;
 using std::cin;
 using std::string;
 using std::vector;
+using std::map;
 
 
 #ifndef FixTHIS
 #define FixTHIS
 
-enum menu_choices { mbalance, mdeposit, mwithdraw, exit };
+// enum menu_choices { mbalance, mdeposit, mwithdraw, exit };
 
-        int attack(int attack)
-        {
-            Attack = attack 
-            return attack;
-        }
+int attack(int weapon, int attack_bonus)
+{
+Attack = attack + weapon
+return Attack;
+}
+
+int health_potion(int health) 
+{
+    5 + health;
+    return health;
+} 
 
 #endif // Fix this. 
+
+
+
+// map<int,myStruct> armor;
+// myStruct t;
+
+// armor.insert(std::map<int,myStruct>::value_type(1, t));  // OK
+
+// armor.insert(1,t);
+// // Error:
+// //   "no instance of overloaded function 'std::map [...]' matches
+// //    the argument list"
+int Armor_Value;
+int hit_Points;
+int armor_Class;
