--- workspace.code-workspace
@@ -1,30 +1,0 @@
-{
-	"folders": [
-		{
-			"name": "Workspace",
-			"path": "."
-		},
-		{
-			"name": "build",
-			"path": "../COSC-Cpp/acc-cosc-1337-fall-2021-Aisett89/build"
-		},
-		{
-			"name": "inc",
-			"path": "../COSC-Cpp/acc-cosc-1337-fall-2021-Aisett89/inc"
-		},
-		{
-			"name": "dnd",
-			"path": "../COSC-Cpp/acc-cosc-1337-fall-2021-Aisett89/dnd"
-		},
-		{
-			"name": "Fun_With_Jake",
-			"path": "../COSC-Cpp/acc-cosc-1337-fall-2021-Aisett89/Fun_With_Jake"
-		}
-	],
-	"settings": {
-		"files.associations": {
-			"string": "cpp",
-			"iosfwd": "cpp"
-		}
-	}
-}
